requirements.txt
amqp==5.3.1
asgiref==3.11.0
billiard==4.2.4
celery==5.6.2
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==6.0.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
kombu==5.6.2
packaging==25.0
pillow==12.1.0
prompt_toolkit==3.0.52
PyJWT==2.10.1
python-dateutil==2.9.0.post0
redis==7.1.0
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
tzlocal==5.3.1
vine==5.1.0
wcwidth==0.2.14
whitenoise==6.11.0
